+++ sentry-android-core/src/test/java/io/sentry/android/core/AndroidOptionsInitializerTest.kt
@@ -281,46 +281,6 @@ class AndroidOptionsInitializerTest {
     }
 
     @Test
-    fun `init should set proguard uuid from properties id on start`() {
-        val assets = mock<AssetManager>()
-
-        whenever(assets.open("sentry-debug-meta.properties")).thenReturn(
-            """
-            io.sentry.ProguardUuids=12ea7a02-46ac-44c0-a5bb-6d1fd9586411
-            """.trimIndent().byteInputStream()
-        )
-
-        fixture.initSut(
-            Bundle(),
-            hasAppContext = false,
-            assets = assets
-        )
-
-        assertNotNull(fixture.sentryOptions.proguardUuid)
-        assertEquals("12ea7a02-46ac-44c0-a5bb-6d1fd9586411", fixture.sentryOptions.proguardUuid)
-    }
-
-    @Test
-    fun `init should set bundle IDs id on start`() {
-        val assets = mock<AssetManager>()
-
-        whenever(assets.open("sentry-debug-meta.properties")).thenReturn(
-            """
-            io.sentry.bundle-ids=12ea7a02-46ac-44c0-a5bb-6d1fd9586411, faa3ab42-b1bd-4659-af8e-1682324aa744
-            """.trimIndent().byteInputStream()
-        )
-
-        fixture.initSut(
-            Bundle(),
-            hasAppContext = false,
-            assets = assets
-        )
-
-        assertTrue(fixture.sentryOptions.bundleIds.size == 2)
-        assertTrue(fixture.sentryOptions.bundleIds.containsAll(listOf("12ea7a02-46ac-44c0-a5bb-6d1fd9586411", "faa3ab42-b1bd-4659-af8e-1682324aa744")))
-    }
-
-    @Test
     fun `init should set Android transport gate`() {
         fixture.initSut()