+++ sentry/src/main/java/io/sentry/Sentry.java
@@ -3,6 +3,8 @@ package io.sentry;
 import io.sentry.cache.EnvelopeCache;
 import io.sentry.cache.IEnvelopeCache;
 import io.sentry.config.PropertiesProviderFactory;
+import io.sentry.internal.debugmeta.NoOpDebugMetaLoader;
+import io.sentry.internal.debugmeta.ResourcesDebugMetaLoader;
 import io.sentry.internal.modules.CompositeModulesLoader;
 import io.sentry.internal.modules.IModulesLoader;
 import io.sentry.internal.modules.ManifestModulesLoader;
@@ -11,6 +13,7 @@ import io.sentry.internal.modules.ResourcesModulesLoader;
 import io.sentry.protocol.SentryId;
 import io.sentry.protocol.User;
 import io.sentry.transport.NoOpEnvelopeCache;
+import io.sentry.util.DebugMetaPropertiesApplier;
 import io.sentry.util.FileUtils;
 import io.sentry.util.thread.IMainThreadChecker;
 import io.sentry.util.thread.MainThreadChecker;
@@ -19,6 +22,7 @@ import java.io.File;
 import java.lang.reflect.InvocationTargetException;
 import java.util.Arrays;
 import java.util.List;
+import java.util.Properties;
 import java.util.concurrent.RejectedExecutionException;
 import org.jetbrains.annotations.ApiStatus;
 import org.jetbrains.annotations.NotNull;
@@ -362,6 +366,12 @@ public final class Sentry {
                   new ResourcesModulesLoader(options.getLogger())),
               options.getLogger()));
     }
+
+    if (options.getDebugMetaLoader() instanceof NoOpDebugMetaLoader) {
+      options.setDebugMetaLoader(new ResourcesDebugMetaLoader(options.getLogger()));
+    }
+    final @Nullable Properties properties = options.getDebugMetaLoader().loadDebugMeta();
+    DebugMetaPropertiesApplier.applyToOptions(options, properties);
 
     final IMainThreadChecker mainThreadChecker = options.getMainThreadChecker();
     // only override the MainThreadChecker if it's not already set by Android